Completely changing my approach
Focus on making a written language that is ugly and awful to use directly, with the understanding that the user will never look at the files themselves directly. The idea here is to make a lisp like language and we can write code in this language to understand its properties and make decisions based on it. Then there are 2 separate apps -- a front-end that displays and edits the source code and a compiler that transforms the source code into executable code, or transpiles it or something. I think separating concerns like this will make things much easier and it will allow me to focus more on actual language design and less on the fine details of the front-end app (this will also simplify the front-end, hopefully, since it will make the front-end exist entirely in text (any edits in the editor can just update the source file and reload it, might be slow but waaaay easier and more reliable to program than a dynamic app with internally stored state)